web/scripts/requirements.txt
requests
python-dotenv
supabase
beautifulsoup4
openai